realizado el merge de todos las ramas que tenia antes para la master

+++ app/controllers/brands_controller.rb
@@ -4,34 +4,37 @@ class BrandsController < ApplicationController
   # GET /brands
   # GET /brands.json
   def index
+    get_all
     @brands = Brand.all
   end
 
-  # GET /brands/1
-  # GET /brands/1.json
+
   def show
+    @types_crops = Brand.find(params[:id])
+    respond_to do |format|
+      format.html { redirect_to  action:"index"}
+      #format.json { render json: @types_crops}
+    end
   end
 
-  # GET /brands/new
+
   def new
+    get_all
     @brand = Brand.new
+    
   end
 
-  # GET /brands/1/edit
+
   def edit
-    @brand = Brand.find(params[:id]);
-    @name = @brand.name
-    @description = @brand.description
+    get_all
   end
 
-  # POST /brands
-  # POST /brands.json
-  def create
+  def create   
     @brand = Brand.new(brand_params)
 
     if @brand.save
       #format.html { redirect_to @product, notice: 'Supply was successfully created.' }
-      render json: { contenido: @brand, location: brand_url(@brands),result: :ok },status: 200
+      render json: { contenido: @brand, location: brand_url(@brand),result: :ok },status: 200
     else
       #format.html { render :new }
       render json:  @brand.errors, status: :unprocessable_entity 
@@ -39,35 +42,38 @@ class BrandsController < ApplicationController
     end
   end
 
-  # PATCH/PUT /brands/1
-  # PATCH/PUT /brands/1.json
   def update
-    @name = params[:brand]["name"];
-    @description = params[:brand]["description"];
-   
-    @brand = Brand.find(params[:id]);
-    @brand.name = @name
-    @brand.description = @description
-    
-    if @brand.save
-      render json: { contenido: @brand,location: "window.location.pathname='#{brands_path}'",result: :ok },status: 200
-    else
-      render json:  @brand.errors, status: :unprocessable_entity 
-    end
+      respond_to do |format|
+        if @brand.update(product_params)
+         format.json { render json: @brand }
+        else
+          format.json { render json: @brands.errors, :status => :unprocessable_entity }
+        end
+      end
+
+  end
+  
+
+  def check_rel(_id)
+    exist_relation = Model.where(:brand_id => _id)
+    return true if exist_relation.count == 0
+    false
   end
 
-  # DELETE /brands/1
-  # DELETE /brands/1.json
   def destroy
-    @brand = Brand.find(params[:id]) 
-    
-    respond_to do |format|
-      if @brand.destroy
-        format.js
-      else
-        format.js
-      end
-    end 
+    @brand        = Brand.find(params[:id]) 
+    if check_rel(params[:id]) 
+        respond_to do |format|
+            if @brand.destroy
+              format.js
+            else
+              render json: { contenido: @brand, location: brand_url(@brand),message: :"no puede ser eliminado" },status: 400
+            end
+        end  
+    else
+      render json:  @brand.errors, status: :bad_request 
+    end
+
   end
 
   private
@@ -80,4 +86,9 @@ class BrandsController < ApplicationController
   def brand_params
     params.require(:brand).permit(:name, :description)
   end
+  
+  def get_all
+    @path = "marcas"
+  end
+
 end
